fix: Remove unnecessary astype conversion for translation columns

+++ src/make_user_translation.py
@@ -15,7 +15,7 @@ def save_user_translation(df, translations, output_file):
     print("DataFrame shape before adding translations:", df.shape)
     print("Translations shape:", translations.shape)
     df = df.reset_index(drop=True)
-    df["translation"] = df.progress_apply(lambda row: translations[row.name], axis=1).astype(np.float32)
+    df["translation"] = df.progress_apply(lambda row: translations[row.name], axis=1)
     df.to_parquet(output_file, index=False)
 
 def main(user_embeddings_file, user_translation_file, model_path):
